Proper target directory support

+++ src/main/java/com/dev9/mvnwatcher/MvnRunner.java
@@ -3,7 +3,6 @@ package com.dev9.mvnwatcher;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -11,7 +10,7 @@ import java.util.concurrent.TimeUnit;
 
 /**
  * Responsible for managing the lifecycle of the mvn execution.
- * <p>
+ * <p/>
  * Monitor started = set dirty flag
  * Directory changed = set dirty flag
  * If dirty flag set, wait 1 second, then stop build if running, and then start build.
@@ -19,10 +18,11 @@ import java.util.concurrent.TimeUnit;
 public class MvnRunner {
 
     Path projectPath;
+    Path targetDirectory;
 
-    public MvnRunner(Path projectPath, List<Task> tasks) {
-
+    public MvnRunner(Path projectPath, Path targetDirectory, List<Task> tasks) {
         this.projectPath = Objects.requireNonNull(projectPath);
+        this.targetDirectory = Objects.requireNonNull(targetDirectory);
         this.tasks = Objects.requireNonNull(tasks);
     }
 
@@ -31,14 +31,12 @@ public class MvnRunner {
     private List<Task> tasks;
 
 
-    private ProcessBuilder getPB(List<String> params, Path directory, File logFile) {
+    private ProcessBuilder getPB(List<String> params, File logFile) {
         ProcessBuilder b = new ProcessBuilder(params);
-        b.directory(directory.toFile());
-
-        Path log = Paths.get(projectPath.toAbsolutePath().toString(), "target", logFile.getAbsolutePath());
+        b.directory(targetDirectory.toFile());
 
         b.redirectErrorStream(true);
-        b.redirectOutput(ProcessBuilder.Redirect.appendTo(log.toFile()));
+        b.redirectOutput(ProcessBuilder.Redirect.appendTo(logFile));
 
         return b;
     }
@@ -47,9 +45,8 @@ public class MvnRunner {
 
         List<ProcessBuilder> configs = new ArrayList<>();
 
-        for(Task task: tasks)
-        {
-            ProcessBuilder pb = getPB(task.toArgList(), projectPath, task.getOutputFile());
+        for (Task task : tasks) {
+            ProcessBuilder pb = getPB(task.toArgList(), task.getOutputFile());
             configs.add(pb);
         }
 
@@ -80,7 +77,6 @@ public class MvnRunner {
 
         @Override
         public void run() {
-            System.out.println("Running hook...");
             monitor.kill();
         }
     }
@@ -107,53 +103,57 @@ public class MvnRunner {
         public void run() {
 
             Exception lastNotifiedException = null;
-            if (lastError != lastNotifiedException) {
-                lastError = lastNotifiedException;
-                if (lastError != null)
-                    System.err.println(lastError.getMessage());
-            }
 
             while (!shutdown) {
+                if (lastError != lastNotifiedException) {
+                    lastError = lastNotifiedException;
+                    if (lastError != null)
+                        System.err.println(lastError.getMessage());
+                }
 
-                try {
-                    dirty = false;
+                if(dirty == true) {
+                    try {
+                        dirty = false;
 
-                    for (int i = 0; i < config.size() - 1; i++) {
+                        for (int i = 0; i < config.size() - 1; i++) {
 
-                        ProcessBuilder pb = config.get(i);
+                            ProcessBuilder pb = config.get(i);
 
-                        System.out.print("Starting " + pb.command().get(0) + "...");
-                        Process p = pb.start();
-                        System.out.println("done. ");
-                        try {
-                            p.waitFor(10, TimeUnit.SECONDS);
-                        } catch (InterruptedException e) {
-                            lastNotifiedException = e;
-                            e.printStackTrace();
-                        } finally {
-                            if (p.isAlive()) {
-                                System.out.println("Destroy Forcibly " + pb.command().get(0));
-                                p.destroyForcibly();
+                            System.out.print("Starting " + pb.command().get(0) + "...");
+
+                            Process p = pb.start();
+
+                            System.out.println("done. ");
+                            try {
+                                p.waitFor(10, TimeUnit.SECONDS);
+                            } catch (InterruptedException e) {
+                                lastNotifiedException = e;
+                                e.printStackTrace();
+                            } finally {
+                                if (p.isAlive()) {
+                                    System.out.println("Destroy Forcibly " + pb.command().get(0));
+                                    p.destroyForcibly();
+                                }
                             }
                         }
+
+                        ProcessBuilder finalConfig = config.get((config.size() - 1));
+
+                        if (watchedProcess == null) {
+                            System.out.print("Starting process (a) " + finalConfig.command().get(0) + "...");
+                            watchedProcess = finalConfig.start();
+                            System.out.println("ready. " + watchedProcess.toString());
+
+                        } else if (!watchedProcess.isAlive()) {
+                            System.out.print("Starting process (b) " + finalConfig.command().get(0) + "...");
+                            watchedProcess = finalConfig.start();
+                            System.out.println("ready. " + watchedProcess.toString());
+                        }
+
+                    } catch (IOException e) {
+                        lastError = e;
+                        e.printStackTrace();
                     }
-
-                    ProcessBuilder finalConfig = config.get((config.size() - 1));
-
-                    if (watchedProcess == null) {
-                        System.out.print("Starting process (a) " + finalConfig.command().get(0) + "...");
-                        watchedProcess = finalConfig.start();
-                        System.out.println("ready. " + watchedProcess.toString());
-
-                    } else if (!watchedProcess.isAlive()) {
-                        System.out.print("Starting process (b) " + finalConfig.command().get(0) + "...");
-                        watchedProcess = finalConfig.start();
-                        System.out.println("ready. " + watchedProcess.toString());
-                    }
-
-                } catch (IOException e) {
-                    lastError = e;
-                    e.printStackTrace();
                 }
 
                 while (dirty == false && watchedProcess.isAlive()) {
